Cast bytes in read buffer to 64bit before bit shifting so sign saturation doesnt mess up the value (0xcc0780009f609d0b instead of 0xffffffff9f609d0b)
git-svn-id: svn://localhost/svn/toolchain/trunk@2810 2adf4935-75e9-4407-93b2-34b21a1880a4

+++ jtag/src/cmd/bfin.c
@@ -209,7 +209,8 @@ cmd_bfin_run( chain_t *chain, char *params[] )
 			    goto execute_cleanup;
 
 			  n = (n << 32)
-			    | (raw_insn[0] << 16) | (raw_insn[1] << 24)
+			    | ((uint64_t)raw_insn[0] << 16)
+			    | ((uint64_t)raw_insn[1] << 24)
 			    | raw_insn[2] | (raw_insn[3] << 8);
 			}
